separate user query from the root query
