+++ api/graphql_route/graph.user.js
@@ -1,8 +1,8 @@
 const graphqlHTTP = require('express-graphql');
-const schema = require('../graphql_schema/scheme.user.graphql');
+const schema = require('../graphql_schema/graphql.RootQuery.js');
 
 module.exports = graphqlHTTP({
- schema,
- // rootValue: root,
- graphiql: true,
+  schema,
+  // rootValue: root,
+  graphiql: true,
 });
